Added content
Added images for the carousel.
Added info section with new icons.
Added icons to asset folder.
Added a bunch of comments with notes in them to be aware of.

Next step is to add the Sign Up area with the button centered inside the image below the info section from old Fantas-e project